requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
python-dotenv
nest_asyncio